pyproject.toml
[project]
authors = [
  { name = "GalacticDynamics", email = "nstarman@users.noreply.github.com" },
  { name = "Nathaniel Starkman", email = "nstarman@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dependencies = [
  "flax>=0.10.0",
  "jax>=0.7.2",
  "jax-tqdm>=0.4.0",
  "jaxtyping>=0.3.5",
  "optax>=0.2.4",
  "optional-dependencies>=0.4.0",
  "plum-dispatch>=2.6.1",
  "unxt>=1.10.0",
  "zeroth>=1.0.1",
]
description = "Nearest Neighbors with Momentum (kNN+p) algorithm for ordering phase-space observations"
dynamic = ["version"]
license.file = "LICENSE"
name = "knnp"
readme = "README.md"
requires-python = ">=3.12"

[project.urls]
"Bug Tracker" = "https://github.com/GalacticDynamics/nearest-neighbours-with-momentum/issues"
Changelog     = "https://github.com/GalacticDynamics/nearest-neighbours-with-momentum/releases"
Homepage      = "https://github.com/GalacticDynamics/nearest-neighbours-with-momentum"


[build-system]
build-backend = "hatchling.build"
requires      = ["hatch-vcs", "hatchling"]

[dependency-groups]
dev = [
  "ipykernel>=7.1.0",
  "matplotlib>=3.10.8",
  "pytest>=9.0.2",
  "sybil[pytest]>=8.0.0",
]
docs = [
  "myst-parser>=4.0.0",
  "pytz>=2024.2",
  "sphinx>=8.1.3",
  "sphinx-autobuild>=2024.10.3",
  "sphinx-autodoc-typehints>=2.6.1",
  "sphinx-book-theme>=1.1.3",
  "sphinx-copybutton>=0.5.2",
  "sphinx-design>=0.6.1",
  "sphinx-prompt>=1.9.0",
  "sphinx-togglebutton>=0.3.2",
  "sphinxcontrib-bibtex>=2.6.3",
  "sphinxext-opengraph>=0.9.2",
]
lint = ["cz-conventional-gitmoji>=0.6.1", "pre-commit>=4.1.0", "pylint>=3.3.8"]


[tool.hatch]
build.hooks.vcs.version-file = "src/knnp/_version.py"
build.hooks.vcs.version-file-template = """\
version: str = {version!r}
version_tuple: tuple[int, int, int] | tuple[int, int, int, str, str]
version_tuple = {version_tuple!r}
"""
version.source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"


[tool.codespell]
ignore-words-list = "leary"
skip              = ["uv.lock"]


[tool.commitizen]
name = "cz_gitmoji"


[tool.ruff]
[tool.ruff.lint]
extend-select = ["ALL"]
ignore = [
  "A002",    # Argument is shadowing a Python builtin
  "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
  "ARG001",  # Unused function argument
  "COM812",  # <- for ruff.format
  "D103",    # Missing docstring in public function
  "D105",    # Missing docstring in magic method
  "D107",    # Missing docstring in __init__
  "D203",    # 1 blank line required before class docstring
  "D213",    # Multi-line docstring summary should start at the second line
  "ERA001",  # Commented out code
  "F722",    # Syntax error in forward annotation <- jaxtyping
  "F811",    # redefinition of unused '...'  <- plum-dispatch
  "F821",    # undefined name '...'  <- jaxtyping
  "FIX002",  # Line contains TODO
  "ISC001",  # Conflicts with formatter
  "N806",    # Variable in function should be lowercase
  "PD",      # Pandas
  "PLR09",   # Too many <...>
  "PLR2004", # Magic value used in comparison
  "PYI041",  # Use `complex` instead of `int | complex` <- plum is more strict
  "RUF022",  # `__all__` is not sorted
  "TD002",   # Missing author in TODO
  "TD003",   # Missing issue link on the line following this TODO
]

[tool.ruff.lint.per-file-ignores]
"**tests/**" = [
  "ANN",
  "E731",
  "INP001",
  "PLC0415",
  "RUF059",
  "S101",
  "SLF001",
  "T20",
]
"__init__.py" = ["F403"]
"conftest.py" = ["ANN", "PLC0415"]
"docs/conf.py" = ["A001", "INP001"]
"noxfile.py" = ["A001", "T20"]
"src/knnp/_src/algorithm.py" = [
  "FBT003", # Boolean literal in function call is clearer than named constant
  "PYI025", # Set from collections.abc is clear in context
  "RUF059", # Unpacked tuple elements required for JAX while loop state
]
"src/knnp/_src/autoencoder.py" = [
  "ANN001", # Missing type annotations for nested loss functions
  "ANN202", # Missing return type for nested loss functions
  "C901",   # train_autoencoder is complex but well-structured
  "RUF003", # Greek letters are intentional in math code
  "RUF059", # Unpacked variable 'keys' used for structure
]
"src/knnp/_src/phasespace.py" = [
  "UP040", # TypeAlias is more explicit for Python 3.12
]

[tool.ruff.lint.flake8-import-conventions.aliases]
equinox                 = "eqx"
"hypothesis.strategies" = "st"
"jax.numpy"             = "jnp"
knnp                    = "kp"

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
known-first-party = [
  "dataclassish",
  "is_annotated",
  "optional_dependencies",
  "quaxed",
  "unxt",
  "xmmutablemap",
  "zeroth",
]
known-local-folder = ["knnp"]

[tool.ruff.format]
docstring-code-format = true


[tool.coverage]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
  'if TYPE_CHECKING:',
]
run.branch = true
run.source = ["knnp"]


[tool.pylint]
ignore-paths = [".*/_version.py"]
messages_control.disable = [
  "cyclic-import",
  "design",
  "duplicate-code",                # Expected for public re-exports
  "fixme",
  "abstract-method",
  "function-redefined",            # plum-dispatch
  "invalid-name",                  # handled by ruff
  "line-too-long",
  "missing-function-docstring",
  "missing-kwoa",                  # plum-dispatch
  "missing-module-docstring",
  "no-value-for-parameter",        # plum-dispatch
  "not-callable",                  # handled by mypy
  "property-with-parameters",      # handled by ruff
  "protected-access",              # handled by ruff
  "redefined-builtin",             # handled by ruff
  "too-many-function-args",        # plum-dispatch
  "unexpected-keyword-arg",        # plum-dispatch
  "unnecessary-lambda-assignment", # Intentional for clarity in some cases
  "unused-argument",               # handled by ruff
  "unused-import",                 # handled by ruff
  "unused-variable",               # handled by ruff
  "unused-wildcard-import",        # handled by ruff
  "wildcard-import",               # handled by ruff
  "wrong-import-order",            # handled by ruff
  "wrong-import-position",
  "W0221",                         # Arguments in function call differ from overridden method
]
py-version = "3.12"
reports.output-format = "colorized"
similarities.ignore-imports = "yes"

[tool.pylint."FORMAT"]
max-module-lines = 2000


[tool.pytest.ini_options]
addopts = ["--showlocals", "--strict-config", "--strict-markers", "-ra"]
filterwarnings = [
  "error",
  # pytest
  "ignore:The hookimpl.*uses old-style configuration options:pytest.PytestDeprecationWarning",
  # jax
  "ignore:jax\\.core\\.pp_eqn_rules is deprecated:DeprecationWarning",
  "ignore:jax\\.core\\.Primitive is deprecated:DeprecationWarning",
  "ignore:jax\\.interpreters\\.batching\\.NotMapped is deprecated:DeprecationWarning",
  "ignore:The numpy\\.array_function protocol:DeprecationWarning",
  "ignore:numpy\\.ndarray size changed:RuntimeWarning",
  # matplotlib
  "ignore::DeprecationWarning:matplotlib",
  # numpy (platform-specific longdouble probe warning during import)
  "ignore:Signature .* for <class 'numpy\\.longdouble'> does not match any known type:UserWarning:numpy\\._core\\.getlimits",
]
log_level = "INFO"
minversion = "8.3"
norecursedirs = [".*", "__pycache__", "docs/_build"]
testpaths = ["README.md", "docs", "src/", "tests/"]
xfail_strict = true
